refactor(types): eliminate service type re-exports and direct all imports to types library (X-Tracker #497)
- Removed all types re-exports from core service modules (\countries.ts\, \	agCatalog.ts\, \mapMentor.ts\, \mentors.ts\) to enforce complete encapsulation.
- Updated all consumer hooks, helper utilities, mock data templates, pages, and components to import types directly from \@/types/*\ instead of \@/services/*\.
- Cleared out unused imports inside service files to keep them warning-free and pristine.
- Verified that the full test suite and production build pass with 100% success.
