+++ src/services/profile/tagCatalog.test.ts
@@ -2,8 +2,8 @@ import { fromPartial } from '@total-typescript/shoehorn';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { apiClient } from '@/lib/apiClient';
+import type { TagCatalogsByBucket, TagCatalogsVO } from '@/types/tagCatalog';
 
-import type { TagCatalogsByBucket, TagCatalogsVO } from './tagCatalog';
 import {
   buildTagLabelMap,
   EMPTY_TAG_CATALOGS,
